flowspace/backend/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
sqlmodel==0.0.8
python-multipart==0.0.6
aiofiles==23.1.0
httpx==0.24.1
openai==1.32.0
